Update faq.md - Fix table rendering issue
Adding a new line between the table text.  This is just a guess based on this line being the only thing that is different from the existing table that works in the same document and the relative changes made to the table that is not rendering correctly.  Fingers crossed! It's worth noting that it does render correctly in the GitHub preview, so GitHub is either more forgiving of incorrect Markdown syntax or it's just rendered differently in the Jekyll context,